Changed from floats to doubles and added PhantomHapticsDevice
git-svn-id: https://www.h3d.org:8090/HAPI/trunk@13 f896c5a5-350d-0410-967e-bc877d8f5295

+++ src/HAPIHapticShape.h
@@ -49,7 +49,7 @@ namespace H3D {
     /// Constructor.
    HAPIHapticShape( void *_userdata,
                      HAPISurfaceObject *_surface,
-                     const H3D::ArithmeticTypes::Matrix4f & _transform ):
+                     const H3D::ArithmeticTypes::Matrix4d & _transform ):
       HAPIHapticObject( _transform ),
       surface( _surface ),
       userdata( _userdata ) {}
